requirements.txt
feedparser
pyexiv2
Pillow
PyQt5
PyPDF2
webcolors
python-magic
requests
mutagen
matplotlib
